+++ src/space/Positionable.java
@@ -15,6 +15,6 @@ public interface Positionable {
 
     public Point2d getPos();
     public ReadOnlyObjectProperty<Point2d> posProperty();
-    public Point2d getPositionIn(long seconds);
+    public Point2d getPositionIn(double seconds);
 
 }
